requirements.txt
alembic==1.13.1
altair==5.3.0
asn1crypto==1.5.1
asttokens==2.4.1
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cryptography==41.0.7
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
distlib==0.3.8
dnspython==2.6.1
docopt==0.6.2
dominate==2.5.1
executing==2.0.1
filelock==3.13.1
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.0.5
fonttools==4.51.0
fsspec==2024.5.0
fuzzywuzzy==0.18.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
huggingface-hub==0.23.2
idna==3.7
intel-openmp==2021.4.0
ipykernel==6.29.4
ipython==8.23.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
Levenshtein==0.25.1
Mako==1.1.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
mysql-client==0.0.1
mysql-connector==2.2.9
mysqlclient==2.2.1
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
np==1.0.2
numpy==1.26.4
packaging==23.2
pandas==2.2.2
parso==0.8.4
pillow==10.3.0
pipenv==2023.12.1
pipreqs==0.4.13
platformdirs==3.11.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pyarrow==15.0.2
pycparser==2.22
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
pymongo==4.6.3
PyMySQL==0.9.3
pyOpenSSL==23.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-editor==1.0.4
python-Levenshtein==0.25.1
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.2
rapidfuzz==3.9.1
referencing==0.34.0
regex==2024.4.28
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
setuptools==69.0.3
six==1.16.0
smmap==5.0.1
snowflake-connector-python==3.7.0
snowflake-sqlalchemy==1.5.1
sonopy==0.1.2
sortedcontainers==2.4.0
SQLAlchemy==1.4.51
stack-data==0.6.3
streamlit==1.33.0
sympy==1.12
tbb==2021.12.0
tenacity==8.2.3
textblob==0.18.0.post0
threadpoolctl==3.5.0
tokenizers==0.19.1
toml==0.10.2
tomlkit==0.12.3
toolz==0.12.1
torch==2.3.0+cu118
torchaudio==2.3.0+cu118
torchvision==0.18.0+cu118
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
transformers==4.41.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
vader==0.0.3
vaderSentiment==3.3.2
virtualenv==20.25.0
visitor==0.1.3
watchdog==4.0.0
wcwidth==0.2.13
Werkzeug==3.0.1
WTForms==2.2.1
xgboost==2.0.3
yarg==0.1.9
